requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
ghasedakpack==0.1.13
idna==3.4
inflection==0.5.1
packaging==23.1
PyJWT==2.8.0
pymongo==3.12.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
sqlparse==0.2.4
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
